Move peer.go to internal/entity
